requirements.txt
tweepy >= 3.8.0
requests >= 2.21.0
configparser
Image
pillow >= 4